+++ Code/Commando/dlgmpslaveservers.h
@@ -53,7 +53,7 @@ class SlaveServerDialogClass : public MenuDialogClass
 		SlaveServerDialogClass(void);
 
 		void On_Init_Dialog(void) override;
-		void On_Command(int ctrl_id, int mesage_id, DWORD param) override;
+		void On_Command(int ctrl_id, int mesage_id, unsigned int param) override;
 		void On_Destroy(void) override;
 		void Load_Settings(int slavenum);
 		static void Set_Slave_Settings(StringClass *file_name);
@@ -61,12 +61,12 @@ class SlaveServerDialogClass : public MenuDialogClass
 
 	private:
 
-		static unsigned long EnableIDs[MAX_SLAVES];
-		static unsigned long NickIDs[MAX_SLAVES];
-		static unsigned long PassIDs[MAX_SLAVES];
-		static unsigned long SerialIDs[MAX_SLAVES];
-		static unsigned long PortIDs[MAX_SLAVES];
-		static unsigned long SettingsButtons[MAX_SLAVES];
+		static unsigned int EnableIDs[MAX_SLAVES];
+		static unsigned int NickIDs[MAX_SLAVES];
+		static unsigned int PassIDs[MAX_SLAVES];
+		static unsigned int SerialIDs[MAX_SLAVES];
+		static unsigned int PortIDs[MAX_SLAVES];
+		static unsigned int SettingsButtons[MAX_SLAVES];
 
 		static char ServerSettingsFileNames[MAX_SLAVES][MAX_PATH];
